// src/app/models/accountreceivable.model.ts
export interface accountreceivableData {
    pledgeId: number;
    country: string;
    accountNo: number;
    bankName: string;
    startDate: Date;
    endDate: Date;
    amount: number;
    installments: number;
    paymentPeriod: number;
    donorId: number;
    donorName:string;
    pledgeFundType: string;
    }